+++ CHANGES.md
@@ -11,6 +11,8 @@
 
 ## develop
 
+## 2022.4.0
+
 - [CHANGE] `--hatch-rate` を `--vcs-hatch-rate` に変更
     - @melpon
 - [ADD] `--instance-hatch-rate` を追加
@@ -22,8 +24,6 @@
 
 - [UPDATE] WebRTC を m102.5005.7.6 に上げる
     - @melpon @voluntas
-- [UPDATE] WebRTC を m102.5005.7.1 に上げる
-    - @melpon
 - [UPDATE] asmjit と blend2d を最新版に上げる
     - @voluntas @melpon
 - [UPDATE] `cmake` を `3.23.2` に上げる